problem api 생성, problem genenrator 배포, problem api 배포
