+++ .gitignore
@@ -99,6 +99,10 @@ infrastructure/cognito/post_confirmation_lambda.zip
 .cache/
 
 infrastructure/api/lambda_zips/*.zip
+infrastructure/problems-api/lambda_zips/*.zip
+problem_generator_v2_lambda.zip
+problem_generator_deps.zip
 
 # cursor
 .cursor/
+
