+++ backend/lambdas/problem-generator-v2/requirements.txt
@@ -1,5 +1,3 @@
-boto3
 langchain
-langchain-core
-langchain-aws 
+
 langchain-google-genai
