+++ frontend/.env.local.sample
@@ -14,3 +14,4 @@ NEXT_PUBLIC_PROBLEM_GENERATION_API_BASE_URL=YOUR_API_GATEWAY_INVOKE_URL # e.g., 
 
 NEXT_PUBLIC_CHATBOT_API_ENDPOINT=YOUR_CHATBOT_API_ENDPOINT # cloudfront_distribution_domain_name
 
+NEXT_PUBLIC_PROBLEM_API_BASE_URL=YOUR_API_GATEWAY_INVOKE_URL # e.g., https://xxxxxxxxx.execute-api.ap-northeast-2.amazonaws.com/stage
