+++ frontend/src/components/Chatbot.tsx
@@ -15,7 +15,7 @@ import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { oneLight } from "react-syntax-highlighter/dist/esm/styles/prism";
-import type { ProblemDetailAPI } from "@/api/dummy/generateProblemApi";
+import type { ProblemDetailAPI } from "@/api/generateProblemApi";
 
 // Define message structure - Use 'model' for AI role
 interface ChatMessage {
