+++ infrastructure/problem-generator-v2/outputs.tf
@@ -7,3 +7,13 @@ output "cloudfront_distribution_domain" {
   description = "The domain name of the CloudFront distribution for Problem Generator V2"
   value       = aws_cloudfront_distribution.problem_generator_v2.domain_name
 }
+
+output "problems_table_name" {
+  description = "The name of the DynamoDB table for storing generated problems"
+  value       = aws_dynamodb_table.problems_table.name
+}
+
+output "problems_table_arn" {
+  description = "The ARN of the DynamoDB table for storing generated problems"
+  value       = aws_dynamodb_table.problems_table.arn
+}
