+++ infrastructure/problem-generator-v2/terraform.auto.tfvars.sample
@@ -1,0 +1,1 @@
+google_ai_api_key     = "YOUR_GOOGLE_AI_API_KEY"
